refactor: combine devmode indicators
